requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
boto3==1.28.25
psycopg2-binary==2.9.7
sqlalchemy==2.0.20
